@@ -35,40 +35,41 @@
     "reflect-metadata": "^0.1.13"
   },
   "devDependencies": {
-    "@angular-devkit/core": "8.0.0",
-    "@angular-devkit/schematics": "8.0.0",
-    "@angular/common": "9.1.13",
-    "@angular/compiler": "9.1.13",
-    "@angular/compiler-cli": "9.1.13",
-    "@angular/core": "9.1.13",
-    "@angular/platform-browser": "9.1.13",
-    "@angular/platform-browser-dynamic": "9.1.13",
-    "@schematics/angular": "8.0.0",
+    "@angular-devkit/core": "^13.2.2",
+    "@angular-devkit/schematics": "^13.2.2",
+    "@angular/common": "^13.2.1",
+    "@angular/compiler": "^13.2.1",
+    "@angular/compiler-cli": "^13.2.1",
+    "@angular/core": "^13.2.1",
+    "@angular/platform-browser": "^13.2.1",
+    "@angular/platform-browser-dynamic": "^13.2.1",
+    "@schematics/angular": "^13.2.2",
     "@types/jasmine": "^3.3.16",
     "@types/sinon": "^7.0.13",
     "copyfiles": "^2.1.1",
     "core-js": "^3.1.4",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
+    "jasmine-core": "~4.0.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "jsonc-parser": "^3.0.0",
+    "karma": "^6.3.15",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "^2.0.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-jasmine": "^4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-typescript": "^4.1.1",
-    "karma-webpack": "^4.0.2",
-    "ng-packagr": "^9.1.3",
-    "rimraf": "^2.6.3",
-    "rxjs": "6.6.7",
-    "ts-loader": "^6.0.4",
-    "ts-node": "~7.0.0",
+    "karma-typescript": "^5.0.0",
+    "karma-webpack": "^5.0.0",
+    "ng-packagr": "^13.2.1",
+    "rimraf": "^3.0.2",
+    "rxjs": "^7.5.2",
+    "ts-loader": "^9.2.6",
+    "ts-node": "^10.4.0",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "tslib": "1.10.0",
-    "tslint": "^5.11.0",
-    "typescript": "3.8.3",
-    "webpack": "^4.38.0",
-    "zone.js": "^0.10.3"
+    "tslib": "^2.0.0",
+    "tslint": "^6.1.3",
+    "typescript": "4.5.5",
+    "webpack": "^5.68.0",
+    "zone.js": "^0.11.4"
   }
 }
